requirements.txt
streamlit==1.44.1
pandas==2.2.3
plotly==6.0.1
scipy==1.15.2
python-dotenv==1.0.1
